Move core package into separate folder
