+++ cmd/openem-ingestor-app/main.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"log"
 
-	core "github.com/SwissOpenEM/Ingestor/internal"
+	core "github.com/SwissOpenEM/Ingestor/internal/core"
 
 	"github.com/spf13/viper"
 	"github.com/wailsapp/wails/v2"
