+++ internal/core/progressnotifier.go
